Add files via upload
The menu.r is the one where if you run the couple of lines before where it calls the login function it will work when run by itself but it doesn't work when you run swirl as a whole. The other file is when we tried to use a global variable but it would keep giving us an error saying that we couldn't change the variable.